tests/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
fcache==0.4.7
flatbuffers==1.12
gast==0.4.0
google-api-core==1.31.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
paramiko==2.7.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
storage==0.0.4.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-serving-api==2.5.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
absl-py==0.13.0
astunparse==1.6.3
attrs==21.2.0
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
cryptography==3.4.7
flatbuffers==1.12
gast==0.4.0
google-api-core==1.31.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
iniconfig==1.1.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
paramiko==2.7.2
pathy==0.6.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
smart-open==5.2.0
storage==0.0.4.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-serving-api==2.5.1
termcolor==1.1.0
toml==0.10.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
pytest-clarity==1.0.1
pytest-instafail==0.4.2
